Near-surface diagnostics in BEP+BEM redefined (A.Martilli)
In module_surface_driver.F near-surface diagnostics (e.g.T2) set to be
equal to the weighted average of the vaules over the urban fraction
(e.g. T2G_URB2D) and vegetation fraction (e.g. T2VEG_URB2D).

NOTE: TCAN_URB2D as it is now is basically the temperature of the
lowest model level - not the temperature of the urban canyon. Defintion
of this TCAN_URB2D is opened to be redefined in the future if needed.
SLUCM is having a different approct to defind this temperature.